requirements.txt
fastapi
uvicorn
openai
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
PyJWT
asyncpg
